List the O_LARGEFILE flag on 64bit
When on a 64bit userspace, O_LARGEFILE is set explicitly whether you
pass it to open or not.

That flag is always set in the processes fdinfo files on 64bit. At the
same time on 64bit the O_LARGEFILE flag is defined as 0, so never gets
printed by fdflags on 64bit.

So we add a check if O_LARGEFILE is defined as 0 (which should signify a
64bit userland) and list the O_LARGEFILE flag, otherwise we only list it
if it is actually set.

Tested on both x86 and x86-64.

On 64bit (the 1 is the O_LARGEFILE flag)

    $ ./fdflags 02100002
    O_LARGEFILE
    O_RDWR
    O_CLOEXEC

On 32bit

    $ ./fdflags 02100002
    O_RDWR
    O_CLOEXEC
    O_LARGEFILE

    $ ./fdflags 02000002
    O_RDWR
    O_CLOEXEC

Note the different order of the O_LARGEFILE flag due to the differnces
in plaves where it's printed out.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>